requirements.txt
python-dotenv
requests
openai
google-genai
google-api-python-client
google-auth
google-auth-oauthlib
celery
redis
Pillow
PyYAML
cryptography
fastapi
jinja2
python-multipart
uvicorn
boto3
supabase
postgrest
itsdangerous
psutil
aiohttp
pyjwt
discord.py
markdown
openai-agents[sqlalchemy]
logfire[fastapi]
stripe
email-validator
greenlet
